refactor(schools): split create/edit/show into hooks and shared sections
Extracts the school forms into create/edit hooks and shared field sections
(school details, principal, address) reused by both pages, plus show-page
cards and table, keeping every file under the size limits. Refs #81.
